refactor(package): payment를 모델로 관리
