+++ src/test/java/lotto/utils/PaymentValidatorTest.java
@@ -1,5 +1,6 @@
 package lotto.utils;
 
+import lotto.model.payment.PaymentValidator;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
